requirements.txt
pandas
numpy
flask
scikit-learn
pytest
requests
tensorflow
keras
requests-oauthlib
Flask-Mail
matplotlib
python-dotenv
pdfplumber
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib